Fixes a %ld format with an int argument.
git-svn-id: http://svn.ghostscript.com/ghostscript/trunk@1910 a1074d23-0009-0410-80fe-cf8c14f379e6

+++ gs/src/gdevpsu.c
@@ -207,11 +207,11 @@ psw_end_file_header(FILE *f)
 void
 psw_end_file(FILE *f, const gx_device *dev,
 	     const gx_device_pswrite_common_t *pdpc, const gs_rect *pbbox,
-             int page_count)
+             int /* should be long */ page_count)
 {
     if (f == NULL)
         return;		/* clients should be more careful */
-    fprintf(f, "%%%%Trailer\n%%%%Pages: %ld\n", page_count);
+    fprintf(f, "%%%%Trailer\n%%%%Pages: %ld\n", (long)page_count);
     if (dev->PageCount > 0 && pdpc->bbox_position != 0) {
 	if (pdpc->bbox_position >= 0) {
 	    long save_pos = ftell(f);
